+++ src/knowledge/prepare-knowledge/markdown/prepareKnowledgeFromMarkdown.ts
@@ -1,20 +1,83 @@
 // import { promptbookStringToJson } from '../../../conversion/promptbookStringToJson';
+import { IVectorData } from 'xyzt';
+import promptbookLibrary from '../../../../promptbook-library/promptbook-library.json';
+import { PromptbookJson } from '../../../_packages/types.index';
+import { normalizeToKebabCase, spaceTrim, string_keyword } from '../../../_packages/utils.index';
+import { assertsExecutionSuccessful } from '../../../execution/assertsExecutionSuccessful';
+import { createPromptbookExecutor } from '../../../execution/createPromptbookExecutor';
 import type { LlmExecutionTools } from '../../../execution/LlmExecutionTools';
+import { createPromptbookLibraryFromSources } from '../../../promptbook-library/constructors/createPromptbookLibraryFromSources';
 import type { KnowledgeJson } from '../../../types/PromptbookJson/KnowledgeJson';
-import type { string_markdown } from '../../../types/typeAliases';
-import { just } from '../../../utils/just';
-// import prepareKnowledgeFromMarkdownStringPromptbook from './prepare-knowledge-from-markdown.ptbk.md';
+import type {
+    string_href,
+    string_markdown,
+    string_markdown_text,
+    string_model_name,
+    string_name,
+} from '../../../types/typeAliases';
 
 export async function prepareKnowledgeFromMarkdown(options: {
-    content: string_markdown;
+    content: string_markdown /* <- TODO: [🖖] Always the file */;
     llmTools: LlmExecutionTools;
 }): Promise<KnowledgeJson> {
     const { content, llmTools } = options;
 
-    // const promptbook =await promptbookStringToJson(prepareKnowledgeFromMarkdownStringPromptbook);
+    const library = await createPromptbookLibraryFromSources(...(promptbookLibrary as Array<PromptbookJson>));
+    const promptbook = library.getPromptbookByUrl(
+        'https://promptbook.studio/promptbook/prepare-knowledge-from-markdown.ptbk.md',
+    );
 
-    just(content);
-    just(llmTools);
+    const executor = createPromptbookExecutor({
+        promptbook,
+        tools: {
+            llm: llmTools,
+            script: [
+                /* <- TODO: Allow to just not define script tools */
+            ],
+        },
+    });
 
-    return [];
+    const result = await executor({ content });
+
+    assertsExecutionSuccessful(result);
+
+    const { outputParameters } = result;
+    const { knowledge: knowledgeRaw } = outputParameters;
+
+    const knowledgeTextPieces = (knowledgeRaw || '').split('\n---\n');
+
+    const knowledge = await Promise.all(
+        knowledgeTextPieces.map(async (knowledgeTextPiece, i) => {
+            // Note: Theese are just default values, they will be overwritten by the actual values:
+            let name: string_name = `piece-${i}`;
+            let title: string_markdown_text = spaceTrim(knowledgeTextPiece.substring(0, 100));
+            const content: string_markdown = spaceTrim(knowledgeTextPiece);
+            const keywords: Array<string_keyword> = [];
+            const index: Array<{ modelName: string_model_name; position: IVectorData }> = [];
+            const sources: Array<{ title: string_markdown_text; href: string_href }> = [];
+
+            try {
+                // TODO: !!!! Summarize name and title from the content
+                title = spaceTrim(knowledgeTextPiece.substring(0, 30));
+                name = normalizeToKebabCase(title);
+
+                // TODO: !!!! Extract keywords via prompt
+                // TODO: !!!! Index through LLM model
+                // TODO: [🖖] !!!! Make system for sources and identification of sources
+            } catch (error) {
+                console.error(error);
+            }
+
+            return {
+                name,
+                title,
+                content,
+                keywords,
+                index,
+                sources,
+            };
+        }),
+    );
+
+    return knowledge;
 }