+++ CHANGELOG.md
@@ -343,6 +343,7 @@ System for management of external knowledge (RAG)
 
 -   Add `MaterialKnowledgePieceJson`
 -   Add `KnowledgeJson`
+-   Add `prepareKnowledgeFromMarkdown` exported from `@promptbook/core`
 -   Change `promptbookStringToJson` to async function
 -   Misc Tool classes not requires options anymore _(like `CallbackInterfaceTools`, `OpenAiExecutionTools`, `AnthropicClaudeExecutionTools`, etc.)_